0.1.13 지혜0.0) 적 동기화 테스트

+++ CyanEngine/Message.h
@@ -4,8 +4,8 @@ struct Message
 {
 	char msgId;
 	int lParam;
-	int mParam;
-	int rParam;
+	float mParam;
+	float rParam;
 };
 
 static std::queue<Message> recvQueue;
